Drop offline migration for SQLite for now
Using the current model is too much of an issue and duplicating two
versions of the involved tables makes the migration scripts overly
verbose, so only support online migrations for SQLite. Since SQLite
basically is a file, pseudo-offline migrations could be handled by
making a copy of the database file.

Demonstrate that offline migration SQL generation is possible for
another dialect by using a mock engine for Postgres in unit tests.

Instead of the stamp command, use the migration API to stamp.