Replace foreign key and enable foreign key constraints
Make the receipt's shop key a normal attribute instead of a foreign key
so that we do not need to add a shop entry for just a key if it has no
other properties for now.

Mark relationships involving products and discounts as passive deletes
because they have "ON DELETE CASCADE" now.

Actually enable foreign keys on SQLite to make them have effect, now
that the constraints seem to be correct as well.

Document the settings, particularly what the foreign keys setting does.