Avoid database event listener problems after test failures
If a database-based test case fails, it will delete its reference to the
database causing it to close, which also removed the listener for sqlite
foreign key pragma setting. This could in some cases cause an error upon
subsequent tests that a deque mutated during iteration, possibly because
the reference deletion was delayed until we establish the new connection
during the next tests.

The event removal is actually not important in normal operation, because
the settings (particularly the database protocol) should not change, but
with more database tests this was causing more noise from failures which
had one cause, so move the event removal to a clear method and use where
needed in tests.