Product inventory: Only use shared fields if set, not in new view step
This makes the view step show the full product metadata in each object
(aside from the shop), instead of category and type being placed in the
common fields (even if these were not part of the pattern fields) when
there are few products. This also applies to the YAML files without the
pattern fields, to keep the portability of the file in case manual edits
are made later on.