Prefer generic match if it has higher specificity than product range
A product has higher specificity if it has more of the three types of
matcher fields (label/price/discount) than another product, or if they
both have the same of these, then it is preferred if it has fewer of
those fields.

During product metadata creation in the new command in item context,
when discounts are not yet known, then the discounts are disregarded
during specificity calculation.