New: Accept metadata when matches are imperfect or duplicate
Product metadata can be added if an item is matched excluding discounts
or duplicate matches are involved. The product metadata can however not
remain standalone because the existing matches will match, but this can
be used to add product ranges or otherwise merge to update metadata. It
is not meant to allow more duplicates, and specificity still applies to
just already-related generic/range products, so it is possible that the
metadata is still rejected in the end. This is mostly to test if adding
another metadata that does not end up matching the item due to discount
differences is not actually added to the inventory, even if part of the
existing range through the merge.