New: Adjust log message for multiple product match
When a new product item matches multiple products that have
non-overlapping discount label matchers, the message should indicate
that it will match one of them assuming a later discount instead of warn
about multiple matches, and adjust log level to follow suit.

Handle overlap between generic product and product range if the discount
matchers are completely the same. This is not as precise as the actual
specificity deduplication that happens to the candidates, but should
lead to clear enough log messages regardless.